pyproject.toml
[tool.poetry]
name = "lola"
version = "0.1.0"
description = "chat bot"
authors = ["Robson Lopes <robson.lopes@edu.unifor.br>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
requests = "^2.28.2"
types-requests = "^2.28.11.15"
pydantic = "^1.10.7"
requests-cache = "^1.0.1"


[tool.poetry.group.dev.dependencies]
pytest = "^7.2.2"
telepot = "^12.7"
isort = "^5.12.0"
black = "^23.1.0"
pycodestyle = "^2.10.0"
flake8 = "^6.0.0"
mypy = "^1.1.1"
pactman = "^2.30.0"
vcrpy = "^4.2.1"
pre-commit = "^3.2.1"

[tool.isort]
force_single_line = true
line_length = 120
lines_between_types = 1
known_third_party = ["urllib3", "pytest"]
sections = ["FUTURE", "STDLIB", "THIRDPARTY", "FIRSTPARTY", "LOCALFOLDER"]

[tool.black]
target_version = ['py39']
include = '\.pyi?$'
exclude = '''
(
  /(
      \.eggs         # exclude a few common directories in the
    | \.git          # root of the project
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | _build
    | buck-out
    | build
    | dist
  )/
)
'''

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
